+++ .github/workflows/test.yml
@@ -9,8 +9,8 @@ jobs:
         strategy:
             matrix:
                 os: [ubuntu-latest,ubuntu-20.04]
-                php: [8.1,8.2]
-                laravel: [^9.0,^10.0]
+                php: [8.1,8.2,8.3]
+                laravel: [^10.0,^11.0]
 
         name: ${{ matrix.os }}, PHP ${{ matrix.php }} / Laravel ${{ matrix.laravel }}
